ref(authz/input) creating reusable interaction input; refactoring authz to use interaction input; moving it into flows dir;
